DRYed up use of the CssHelper and moved it under the DesignSystem namespace
