+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Cleaned up require calls with Zeitwerk
+- DRYed up use of the CssHelper and moved it under the DesignSystem namespace
 
 ## [0.7.0] - 2025-08-22
